requirements.txt
# ──────────────────────────────────────────────
# Multimodal Math Mentor — Python Dependencies
# ──────────────────────────────────────────────

# Core
streamlit>=1.32.0
python-dotenv>=1.0.0

# LLM providers
openai>=1.30.0
anthropic>=0.25.0
groq>=0.5.0
google-genai>=1.0.0

# LangChain ecosystem
langchain>=0.2.0
langchain-openai>=0.1.0
langchain-community>=0.2.0
langchain-text-splitters>=0.2.0

# OCR — Mathpix (primary, cloud API) + EasyOCR (offline fallback)
# Mathpix is accessed via built-in urllib — no extra package required.
# Install requests only if you prefer it:
# requests>=2.31.0
easyocr>=1.7.0

# Speech-to-text
openai-whisper>=20231117

# Vector database
chromadb>=0.5.0

# Embeddings
sentence-transformers>=2.7.0

# Math
sympy>=1.12

# Visualization (Manim)
manim>=0.18.0
imageio-ffmpeg>=0.5.0   # bundled ffmpeg binary for Windows

# Database
# (sqlite3 is built-in)

# Utilities
Pillow>=10.0.0
numpy>=1.26.0
pydantic>=2.5.0
soundfile>=0.12.0

# Testing
pytest>=8.0.0
